fixed errors

+++ src/Routers/taskRouter.js
@@ -7,14 +7,11 @@ const taskApp = new express.Router()
 // Creating Task
 
 taskApp.post('/tasks', auth ,async (req, res) => {
-    const task = new Task({
-        ...req.body,
-        owner : req.user._id
-    })
+    const task = new Task({...req.body, owner : req.user._id})
 
     try {
         const createdtask = await task.save()
-            res.status(201).send(createdtask)
+        res.status(201).send(createdtask)
     }
     catch (error) {
         res.status(400).send(error)
@@ -23,10 +20,13 @@ taskApp.post('/tasks', auth ,async (req, res) => {
 
 // Reading Tasks
 
-taskApp.get('/tasks', async (req, res) => {
+taskApp.get('/tasks', auth , async (req, res) => {
     try {
-        const tasks = await Task.find({})
-            res.send(tasks)
+        const tasks = await Task.find({owner : req.user._id})
+        if (!tasks) {
+            return res.send({error: "No Tasks Found."})
+        }
+        res.send(tasks)
     }
     catch (error) {
         res.status(500).send(error)
@@ -35,13 +35,13 @@ taskApp.get('/tasks', async (req, res) => {
 
 // Reading Singular Task by ID
 
-taskApp.get('/tasks/:id', async (req, res) => {
+taskApp.get('/tasks/:id', auth , async (req, res) => {
     const _id = req.params.id
 
     try {
-        const task = await Task.findById({ _id })
+        const task = await Task.findOne({_id , owner : req.user._id })
             if (!task) {
-                return res.status(404).send()
+                return res.status(404).send({error : "Task Not Found"})
             }
             res.send(task)
     }
@@ -52,7 +52,7 @@ taskApp.get('/tasks/:id', async (req, res) => {
 
 // Updating Tasks
 
-taskApp.patch('/tasks/:id' , async (req , res) => {
+taskApp.patch('/tasks/:id' , auth , async (req , res) => {
     const updates = Object.keys(req.body)
     const allowedUpdates = ["desc" , "completed"]
     const isValidOperation = updates.every( (update) => allowedUpdates.includes(update) )
@@ -61,30 +61,28 @@ taskApp.patch('/tasks/:id' , async (req , res) => {
         return res.status(404).send({error : "Invalid Input"})
     }
     try {
-        const task = await Task.findById(req.params.id)
+        const task = await Task.findById({ _id: req.params.id , owner: req.user._id})
+
+        if (!task) {
+            return res.status(404).send("Task Not Found")
+        }
         updates.forEach((update) => task[update] = req.body[update])
         await task.save()
-
-        // const task = await Task.findByIdAndUpdate(req.params.id , req.body , {new : true , runValidators : true})
-        if (!task) {
-            return res.status(404).send()
-        }
         res.status(201).send(task)
     }
     catch(error) {
         res.status(500).send(error)
-        console.log(error)
     }
 })
 
 // Deleting Tasks 
 
-taskApp.delete('/tasks/:id' , async (req , res) => {
+taskApp.delete('/tasks/:id' , auth , async (req , res) => {
     try {
-        const task = await Task.findByIdAndDelete(req.params.id)
+        const task = await Task.findByIdAndDelete({ _id : req.params.id , owner : req.user._id})
 
         if(!task){
-            return res.status(404).send({error : "Id Not Found"})
+            return res.status(404).send({error : "Task Not Found"})
         }
         res.status(201).send({message : "Task Deleted Successfully"})
     }
